smart-bin-middleman/flask-app/requirements.txt
flask
paho-mqtt
firebase-admin